Allow specific rows to stay at bottom on reversing sort.
Types can define an additional one-argument function, bottom, that
identifies values that should stay at the bottom of the table when the
table's sort order is reversed. A typical use case would be to sort
empty values to the bottom regardless of the direction of sorting.

To keep a consistent sort order, types that define bottom must ensure
that values for which bottom returns true are sorted to the bottom by
the compare function, as well.